+++ packages/core/actors/eventObservable/package.json
@@ -1,0 +1,8 @@
+{
+  "main": "dist/xstate-actors-eventObservable.cjs.js",
+  "module": "dist/xstate-actors-eventObservable.esm.js",
+  "umd:main": "dist/xstate-actors-eventObservable.umd.min.js",
+  "preconstruct": {
+    "umdName": "XStateActorsEventObservable"
+  }
+}
