+++ packages/core/actors/promise/package.json
@@ -1,0 +1,8 @@
+{
+  "main": "dist/xstate-actors-promise.cjs.js",
+  "module": "dist/xstate-actors-promise.esm.js",
+  "umd:main": "dist/xstate-actors-promise.umd.min.js",
+  "preconstruct": {
+    "umdName": "XStateActorsPromise"
+  }
+}
