+++ packages/core/package.json
@@ -10,10 +10,6 @@
       "module": "./dist/xstate.esm.js",
       "default": "./dist/xstate.cjs.js"
     },
-    "./actors": {
-      "module": "./actors/dist/xstate-actors.esm.js",
-      "default": "./actors/dist/xstate-actors.cjs.js"
-    },
     "./guards": {
       "module": "./guards/dist/xstate-guards.esm.js",
       "default": "./guards/dist/xstate-guards.cjs.js"
@@ -25,6 +21,26 @@
     "./dev": {
       "module": "./dev/dist/xstate-dev.esm.js",
       "default": "./dev/dist/xstate-dev.cjs.js"
+    },
+    "./actors": {
+      "module": "./actors/dist/xstate-actors.esm.js",
+      "default": "./actors/dist/xstate-actors.cjs.js"
+    },
+    "./actors/promise": {
+      "module": "./actors/promise/dist/xstate-actors-promise.esm.js",
+      "default": "./actors/promise/dist/xstate-actors-promise.cjs.js"
+    },
+    "./actors/reducer": {
+      "module": "./actors/reducer/dist/xstate-actors-reducer.esm.js",
+      "default": "./actors/reducer/dist/xstate-actors-reducer.cjs.js"
+    },
+    "./actors/callback": {
+      "module": "./actors/callback/dist/xstate-actors-callback.esm.js",
+      "default": "./actors/callback/dist/xstate-actors-callback.cjs.js"
+    },
+    "./actors/observable": {
+      "module": "./actors/observable/dist/xstate-actors-observable.esm.js",
+      "default": "./actors/observable/dist/xstate-actors-observable.cjs.js"
     },
     "./package.json": "./package.json"
   },
@@ -75,7 +91,11 @@
     "entrypoints": [
       "./index.ts",
       "./actions.ts",
-      "./actors.ts",
+      "./actors/index.ts",
+      "./actors/reducer.ts",
+      "./actors/promise.ts",
+      "./actors/observable.ts",
+      "./actors/callback.ts",
       "./guards.ts",
       "./dev/index.ts"
     ]
