+++ packages/core/test/interpreter.test.ts
@@ -18,7 +18,9 @@ import { stop } from '../src/actions/stop';
 import { log } from '../src/actions/log';
 import { isObservable } from '../src/utils';
 import { interval, from } from 'rxjs';
-import { fromCallback, fromObservable, fromPromise } from '../src/actors';
+import { fromObservable } from '../src/actors/observable';
+import { fromPromise } from '../src/actors/promise';
+import { fromCallback } from '../src/actors/callback';
 
 const lightMachine = createMachine({
   id: 'light',
